use module to format boss data
